backend/pyproject.toml
[tool.poetry]
name = "_backend"
version = "0.1.0"
description = ""
authors = ["mikheyev9 <mikheyev9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
tortoise-orm = "^0.21.6"
uvicorn = "^0.30.6"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
aerich = "^0.7.2"
gunicorn = "^23.0.0"

[tool.poetry.scripts]
start-app = "start_app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
